+++ api/provider/aws/service_create.go
@@ -1,0 +1,8 @@
+package aws
+
+import "github.com/quintilesims/layer0/common/models"
+
+func (s *ServiceProvider) Create(req models.CreateServiceRequest) (*models.Service, error) {
+
+	return nil, nil
+}
