+++ api/provider/aws/service_delete.go
@@ -1,0 +1,5 @@
+package aws
+
+func (s *ServiceProvider) Delete(ServiceID string) error {
+	return nil
+}
