+++ api/provider/mock_provider/mock_service.go
@@ -71,3 +71,15 @@ func (_m *MockServiceProvider) Read(_param0 string) (*models.Service, error) {
 func (_mr *_MockServiceProviderRecorder) Read(arg0 interface{}) *gomock.Call {
 	return _mr.mock.ctrl.RecordCall(_mr.mock, "Read", arg0)
 }
+
+// Update mocks base method
+func (m *MockServiceProvider) Update(arg0 models.UpdateServiceRequest) error {
+	ret := m.ctrl.Call(m, "Update", arg0)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// Update indicates an expected call of Update
+func (mr *MockServiceProviderMockRecorder) Update(arg0 interface{}) *gomock.Call {
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Update", reflect.TypeOf((*MockServiceProvider)(nil).Update), arg0)
+}
